+++ CHANGELOG.md
@@ -10,9 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added support for MatrixInput and -Output Gain/Mute to MIDI Bridging
 
 ### Changed
-- 
+- Introduced massive performance improvement regarding configuration change handling (app start&quit, SoundObject/MatrixInput/MatrixOutput batch add&remove, xml config loading)
 
 ### Fixed
+- Fixed xml config file loading
 - Fixed MIDI handling for SoundObject Delay Mode in case three note on/off triggers are assigned
 
 ## [0.3.5] - 2021-07-03